refactor(view): inputView 분리
